@@ -12,15 +12,15 @@
   "private": true,
   "dependencies": {
     "-": "^0.0.1",
-    "@angular/animations": "~13.4.0",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/localize": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/animations": "~14.3.0",
+    "@angular/common": "~14.3.0",
+    "@angular/compiler": "~14.3.0",
+    "@angular/core": "~14.3.0",
+    "@angular/forms": "~14.3.0",
+    "@angular/localize": "~14.3.0",
+    "@angular/platform-browser": "~14.3.0",
+    "@angular/platform-browser-dynamic": "~14.3.0",
+    "@angular/router": "~14.3.0",
     "@ng-bootstrap/ng-bootstrap": "^9.1.3",
     "bootstrap": "^4.5.0",
     "g": "^2.0.1",
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1100.5",
     "@angular/cli": "^11.0.5",
-    "@angular/compiler-cli": "~13.4.0",
+    "@angular/compiler-cli": "~14.3.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
