@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-my-form1',
@@ -8,10 +8,10 @@ import { FormBuilder, FormGroup } from '@angular/forms';
 })
 export class MyForm1Component implements OnInit {
 
-  myForm!: FormGroup;
+  myForm!: UntypedFormGroup;
 
   constructor(
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
   ) { }
 
   ngOnInit(): void {
